add test cases for #875, #874, buildTags, type interface match,fix false test cases bug in #858.
