+++ testdata/inspection/constantExpressionsInConstDeclarations/issue874.go
@@ -1,0 +1,7 @@
+package constantExpressionsInConstDeclarations
+
+type During uint64
+
+const Millisecond During = 1000
+
+const flushDelay = 100 * Millisecond
