+++ testdata/inspection/functionCall/issue812.go
@@ -39,28 +39,3 @@ func good2() {
 	f2(f2([]int{1, 2, 3}...))
 	f2(f11())
 }
-
-/*
-TODO finish it when type system can do it.
-wait for interface isIdentical check and type assignability check.
-func f3(arg ...interface{}) int {
-	return 0
-}
-func good3() {
-	a := []interface{}{}
-	f3(a...)
-}
-
-type t1 string
-
-func f4(arg ...t1) int {
-	return 0
-}
-func bad4() {
-	f4([]string{}...)
-}
-func good4() {
-	a := []t1{}
-	f4(a...)
-}
-*/
