+++ testdata/inspection/functionCall/issue875/p1/p1.go
@@ -1,0 +1,7 @@
+package p1
+
+func F1(name string, mode T1) error {
+	return nil
+}
+
+type T1 uint32
