+++ testdata/inspection/functionCall/issue875/p2/p2.go
@@ -1,0 +1,10 @@
+package p2
+
+import (
+	"issue875/p1"
+)
+
+func f1() {
+	p1.F1("123", p1.T1(0700))
+}
+
