+++ testdata/inspection/redeclare/buildTags/race.go
@@ -1,0 +1,14 @@
+
+// +build race
+
+package buildTags
+
+func raceDisable() {
+	a:=1
+	_=a
+}
+
+func raceEnable() {
+	a:=1
+	_=a
+}
