+++ testdata/inspection/redeclare/buildTags/race0.go
@@ -1,0 +1,10 @@
+
+// +build !race
+
+package buildTags
+
+func raceDisable() {
+}
+
+func raceEnable() {
+}
