+++ testdata/inspection/unresolvedSymbols/issue858/p1/pf2.go
@@ -1,0 +1,5 @@
+package p1
+
+func (t T4)F4(){
+
+}
